Add libcurl SSL support via gnutls
Gnutls is used as it has the best license compatibility with
GNU Scientific Library should we want to enabe the BCFtools
polysomy plugin.

Includes gnutls dependencies gmp and nettle.  In theory it should
be possible to build just with nettle using its built-in mini-gmp
but unfortunately that option seems to be broken.  LibIDN is
disabled to reduce the number of dependencies we have to
include.

Adds a wrapper around __vasprintf_chk.  Unfortunately the checks
involve glibc stdio internals so are difficult to replicate.
Instead we just redirect back to the original vasprintf.