Enable S3 URLs
Requires an HMAC-SHA1 function.  Provide this using nettle, which
is already being built for gnutls (for libcurl).  A small wrapper
is used to give an openssl style interface - not quite the genuine
one, but enough to satisfy the configure script and hfile_s3.c.

Nettle's lower-level interface is used instead of the one provided
by gnutls because the latter is fairly big and expands the
binaries it's linked in to quite a lot.