Add tests for samtools and bcftools binaries.
Copy, or link to, parts of the samtools and bcftools test harnesses
to allow the final binaries to be tested.

Add rule to make a tar file with the test scripts and data in it.
This can be copied to another machine without any of the build
dependencies along with the binaries for testing.  (Perl is
required on the target machine for the tests to work).