Inject CA bundle finding code into libcurl itself
It turns out that libcurl reports an error if the CA bundle
file or CA path given at configure time does not exist and
it tries to use them.  As they may not exist on the system
where the binaries are run, don't try to set them at all.

The code to search for the CA bundle file only ran if the
libcurl wrapper knew it was loading the fallback libcurl.  It
got skipped if LD_LIBRARY_PATH was used to force the fallback
library to be loaded instead of the system one.  To solve this,
the search code it moved into a function annotated with
__attribute__((constructor)) that is included when linking the
fallback libcurl.  That way it gets run however the fallback
library was loaded.