Fix libcurl breakages
Wrong library name in libcurl_wrap.c, now we're not shipping the
libcurl.so symlink.

The function to find the CA files was getting excluded from
the libcurl shared library file for some reason.  Ensure it
goes in by using the --undefined linker flag so it has
to be resolved.